Switch MyHaydAppTests to MyHadhAppTest.kt
